@@ -1,17 +1,17 @@
 name: flutter_nfc_acs
 description: A plugin for communicating with the ACR1255U-J1 BT reader from Advanced Card Systems Ltd.
-version: 0.1.6
-author: Lars Støttrup Nielsen
+version: 0.2.0
+author: nuvoPoint ApS
 homepage: https://www.nuvopoint.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_plugin_android_lifecycle: ^1.0.6
+  flutter_plugin_android_lifecycle: ^2.0.0
 
 dev_dependencies:
   flutter_test:
